+++ CHANGELOG.md
@@ -8,9 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+- NA
+
+## [6.5.3] - 2023-11-17
+
 - Updated [Measure-Summary] synopsis
 - Supported [grep] `-LeaveHeaderAndBoarder` option when used with `-Context` option
 - Fixed [Join2-Object] Synopsis
+- Added [Replace-ForEach] `-OnlyIfPropertyExists` option
 
 ## [6.5.2] - 2023-11-01
 
@@ -559,7 +564,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Transpose-Property]: src/Transpose-Property_function.ps1
 
 
-[unreleased]: https://github.com/btklab/posh-mocks/compare/6.5.2..HEAD
+[unreleased]: https://github.com/btklab/posh-mocks/compare/6.5.3..HEAD
+[6.5.3]: https://github.com/btklab/posh-mocks/releases/tag/6.5.3
 [6.5.2]: https://github.com/btklab/posh-mocks/releases/tag/6.5.2
 [6.5.1]: https://github.com/btklab/posh-mocks/releases/tag/6.5.1
 [6.5.0]: https://github.com/btklab/posh-mocks/releases/tag/6.5.0